requirements-dev.txt
aiohttp>=3.7.4.post0
web3>=5.20.0
pytest>=6.2.4
pytest-asyncio>=0.15.1
pytest-dependency>=0.5.1
marshmallow-dataclass>=8.4.1
setuptools>=42
wheel